Update theme.css
.panel-profile .panel-profile-header .img-profile-header-background { width: 100%; max-height: 552px; }
changed from max-height: 192px; changed to max-height:552px - so that also bigger header images can be processed without accessing the core code. Until now it is impossible to change that small flat narrow header image to a nearly full-size one. Especially for spaces and profiles, it is great to have the ability to insert also a higher image. the max height of 552px has been chosen so that still some data appears below the image on screens 1280 wide. A higher image gives also much better options to choose a regular smartphone or 4x3, 16x9, 16x10 image and then choose parts of that image to be displayed as header image without need of any biffer header modifications. From the feedback with our users here most - actually all - complaint about that way to narrow header image as it is not possible to create a nice header image more or less instantly with a camera or smartphone and that means the users alll need extra tools to achieve a nice looking page. The PHP part pull request which is needed too has already been send. Thanks!